Refactor: Move definição do server para pasta de infra
